% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_server}
\alias{apex_server}
\title{Apex Charts}
\usage{
apex_server(input, output, session, apexc = NULL, base_data)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{apexc}{the input function of an apex chart. Often times a defined reactive function}

\item{base_data}{the underlying apex data for refreshing the plot on the module side render}
}
\description{
An apexchart.js library integration
in shiny to provide interactive charts for PWA
on the server-side of an application
}
