% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apex.R
\name{apex_ui}
\alias{apex_ui}
\title{Apex Charts}
\usage{
apex_ui(id)
}
\arguments{
\item{id}{the namespace id of the apexchart}
}
\description{
An apexchart.js library integration
in shiny to provide interactive charts for PWA
}
