% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spec.R
\name{as_class}
\alias{as_class}
\title{Standard class specifications}
\usage{
as_class(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{A class specification.
\itemize{
\item An R7 class object or class union.
\item An S3 class object, created by \code{s3_class()}.
\item An S4 class object.
\item A base type specified either with its constructor (\code{logical}, \code{integer},
\code{double} etc) or its name (\code{"logical"}, \code{"integer"}, "\code{double}" etc).
}}

\item{arg}{Argument name used when generating errors.}
}
\value{
A standardised class: either \code{NULL}, an R7 class, an R7 union,
as \link{s3_class}, or a S4 class.
}
\description{
This is used as the interface between R7 and R's other OO systems, allowing
you to use R7 classes and methods with base types, informal S3 classes, and
formal S4 classes.
}
