% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_server}
\alias{card_server}
\title{THe value box server for Framework 7}
\usage{
card_server(
  input,
  output,
  session,
  name,
  image = NULL,
  footer = NULL,
  outline = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{name}{The title of the card}

\item{image}{The image used for an expandable card}

\item{footer}{The footer appended to the bottom of the card}

\item{outline}{Outline style of the card}

\item{height}{The complete height of the card}

\item{...}{The other components of the card UI}
}
\description{
The value box creates a server-side function to be handled
by conditional logic at later dates for the parameters
}
