% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card_ui}
\alias{card_ui}
\title{THe Card UI for Framework 7}
\usage{
card_ui(id)
}
\arguments{
\item{id}{is unique ID associated with the button for the UI namespace}
}
\description{
The card creates a server-side function to be handled
by conditional logic at later dates
}
