% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\title{Cast an object from one type to another}
\usage{
cast(from, to, ...)
}
\arguments{
\item{from}{An R7 object to cast.}

\item{to}{An R7 class specification, passed to \code{\link[=as_class]{as_class()}}.}

\item{...}{Other arguments passed to custom \code{cast()} methods.}
}
\description{
\code{cast()} is a non-standard generic: it dispatch on the terminal class of
\code{from} and \code{to} (and unlike normal dispatch, \code{to} is a class, not an object).

\code{cast()} provides an automatic fallback if \code{from} inherits from \code{to}. You
can override this if you need some special behavior other than simply
stripping class.
}
