% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\docType{data}
\name{class_any}
\alias{class_any}
\title{Dispatch on any class}
\usage{
class_any
}
\description{
Use \code{class_any} to register a default method that is called when no other
methods are matched.
}
\examples{
foo <- new_generic("foo", "x")
method(foo, class_numeric) <- function(x) "number"
method(foo, class_any) <- function(x) "fallback"

foo(1)
foo("x")
}
\keyword{datasets}
