% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{class_get}
\alias{class_get}
\title{Retrieve the R7 class from a class specification}
\usage{
class_get(x, unions = FALSE, envir = parent.frame())
}
\arguments{
\item{x}{The name of the R7 class}

\item{unions}{Include unions?}

\item{envir}{The environment to look for the name}
}
\description{
Retrieve the R7 class from a class specification
}
