% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{class_missing}
\alias{class_missing}
\title{Dispatch on a missing argument}
\usage{
class_missing
}
\value{
Sentinel objects used for special types of dispatch.
}
\description{
Use \code{class_missing} to dispatch when no argument value is available.
Omitted arguments with defaults dispatch on the default's class.
Missing values detected by \code{\link[=is.na]{is.na()}} do not trigger \code{class_missing} dispatch.
}
\examples{
foo := new_generic("x")
method(foo, class_numeric) <- function(x) "number"
method(foo, class_missing) <- function(x) "missing"
method(foo, class_any) <- function(x) "fallback"

foo(1)
foo()
foo("")
}
