% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{class_names}
\alias{class_names}
\title{Retrive all of the class names for a class}
\usage{
class_names(object)
}
\arguments{
\item{object}{The R7 object to query}
}
\value{
A character vector of all the class names for a given R7 class.
}
\description{
Retrive all of the class names for a class
}
