% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{class_new}
\alias{class_new}
\title{Define a new R7 class}
\usage{
class_new(
  name,
  parent = r7_object,
  constructor = function(...) object_new(...),
  validator = function(x) NULL,
  properties = list()
)
}
\arguments{
\item{name}{The name of the class}

\item{parent}{The parent class}

\item{constructor}{The constructor function}

\item{validator}{The validation function}
}
\description{
Define a new R7 class
}
