% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{class_union}
\alias{class_union}
\title{Class unions}
\usage{
class_union(...)
}
\description{
A class union represents a list of possible classes. It is used in
properties to allow a property to be one of a set of classes, and in method
dispatch as a convenience for defining a method for multiple classes.
}
