% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date_server}
\alias{date_server}
\title{Framework 7 date picker server-side logic}
\usage{
date_server(
  input,
  output,
  session,
  val = Sys.Date(),
  multiple = T,
  label = "Pick a Date",
  minDate = as.Date("2020-01-01"),
  maxDate = Sys.Date() + 15,
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{val}{The default value set for the date picker UI}

\item{multiple}{Logical. Enable multiple selection of choices}

\item{label}{The label on the date picker UI}

\item{minDate}{Minimum date option for the date picker UI}

\item{maxDate}{Minimum date option for the date picker UI}

\item{...}{Additional parameters for f7DatePicker()}
}
\description{
The server-side function used in
shinyMobile developed shiny applications
}
