% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date_ui}
\alias{date_ui}
\title{Framework 7 Date Picker}
\usage{
date_ui(id, ...)
}
\arguments{
\item{id}{the namespace id of the date picker UI}

\item{...}{additional parameters to uiOutput function}
}
\description{
The framework 7 date Picker input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
