% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dt_server}
\alias{dt_server}
\title{Datatable (DT) Server side}
\usage{
dt_server(
  input,
  output,
  session,
  base_data,
  row_n = T,
  col_names = NULL,
  currency_pattern = "sales|aov|budg\\\\w+",
  pct_pattern = "\\\\\%",
  opts = list(),
  dt_height = NULL,
  s = T,
  escape = F
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{base_data}{A data frame to display within the datatable() function call}

\item{row_n}{Boolean. TRUE or FALSE to include row names (converted to reactive on server side)}

\item{col_names}{Character. Accepts a character array that corresponds to the columns in the data frame}

\item{currency_pattern}{Character. Regular expression for finding the columns that have a currency format}

\item{pct_pattern}{Character. Regular expression for finding the columns that have a percent format}

\item{opts}{List. A list of additional options that are passed to DT::datatable(... options = opts)}

\item{dt_height}{Character. A pixel, em or css height adjustment for the datatable rendered}

\item{s}{Boolean. Toggle whether to have a server side or client side rendering of the datatable's data (Default = T)}

\item{escape}{Boolean. Toggle whether to escape HTML rendering (Default = F)}
}
\description{
The DT server-side functionality
used in a Framework 7 application
}
