% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dt_ui}
\alias{dt_ui}
\title{Data Table (DT) UI}
\usage{
dt_ui(id, ...)
}
\arguments{
\item{id}{the namespace id of the radio UI}

\item{...}{parameters passed to the uiOutput}
}
\description{
The Datatable UI for the client side
used in a Framework 7 shiny application
}
