% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7li.R
\name{f7li_server}
\alias{f7li_server}
\title{Framework 7 list item}
\usage{
f7li_server(
  input,
  output,
  session,
  title = NULL,
  subtitle = NULL,
  header = NULL,
  footer = NULL,
  href = NULL,
  media = NULL,
  right = NULL,
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{title}{Title used in the list item}

\item{subtitle}{The subtitle of the list item}

\item{header}{Header component of the list item}

\item{footer}{Footer component of the list item}

\item{href}{hyperlink reference for the list item group}

\item{media}{The reference for the framework 7 icon used in the list item}

\item{right}{The component on the righthand side of the list item}

\item{...}{Item text}
}
\description{
The server-side integration of the framework 7
list items for use in an f7listgroup function call
}
