% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7li.R
\name{f7li_ui}
\alias{f7li_ui}
\title{Framework 7 list item}
\usage{
f7li_ui(id)
}
\arguments{
\item{id}{the namespace id of the list items}
}
\description{
A Framework 7 list item function
to conveniently build list group
}
