% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7SmartSelect.R
\name{f7SmartSelect}
\alias{f7SmartSelect}
\title{f7SmartSelect updated from shinyMobile error in NS}
\usage{
f7SmartSelect(
  inputId,
  label,
  choices,
  selected = NULL,
  openIn = c("page", "sheet", "popup", "popover"),
  searchbar = TRUE,
  multiple = FALSE,
  maxlength = NULL,
  virtualList = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The id of the input object.}

\item{label}{Select input label.}

\item{choices}{The new choices.}

\item{selected}{The new value for the input.}

\item{openIn}{Smart select type: either \code{c("sheet", "popup", "popover")}.
Note that the search bar is only available when the type is popup.}

\item{searchbar}{Whether to enable the search bar. TRUE by default.}

\item{multiple}{Whether to allow multiple values.}

\item{maxlength}{Maximum items to select when multiple is TRUE.}

\item{virtualList}{Enable Virtual List for smart select if your select has a lot
of options. Default to FALSE.}

\item{...}{Parameters used to update the smart select,
use same arguments as in \code{\link[shinyMobile]{f7SmartSelect}}.}
}
\description{
f7SmartSelect updated from shinyMobile error in NS
}
\seealso{
\link{f7SmartSelect}
}
