% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{find_call}
\alias{find_call}
\alias{is_ns_call}
\title{Find a call in expression}
\usage{
find_call(x, name, ns = NULL)

is_ns_call(x, name, ns = NULL)
}
\arguments{
\item{x}{An language object in which to search for a particular function.}

\item{name}{A quoted function name to search for. If \code{ns} is provided, calls
to either the plain function name and namespace-qualified function name are
considered.}

\item{ns}{A namespace for the call. If \code{NULL} (the default), the namespace is
unused when matching the call. If a string, the call may also match a
namespace-qualified call.}
}
\value{
\code{call} object if found, or \code{NULL} otherwise.

a \code{logical} value indicating whether \code{x} is a call to \code{ns::name}
}
\description{
Find a call (namespaced or plain) in an language object.
}
\keyword{internal}
