% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{generic_generate}
\alias{generic_generate}
\title{Generate the body of a generic function}
\usage{
generic_generate(name, signature, envir = parent.frame())
}
\arguments{
\item{name}{of the generic}

\item{signature}{signature of the generic}

\item{envir}{environment to use as the enclosing environment of the generated generic}
}
\description{
This is used as the default to \link{new_generic}
}
