% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\alias{method_register}
\title{Retrieve or register an r7 method for a generic}
\usage{
method(generic, signature, envir = parent.frame())

method_register(generic, signature, value, envir = parent.frame())
}
\arguments{
\item{generic}{The generic to retrieve or register}

\item{signature}{The method signature}

\item{envir}{The environment to lookup the generic in.}

\item{value}{The new function to use for the method.}
}
\description{
Retrieve or register an r7 method for a generic
}
