% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method_next}
\alias{method_next}
\title{Retrieve the next applicable method after the current one}
\usage{
method_next(generic, signature)
}
\arguments{
\item{generic}{The generic to retrieve or register}

\item{signature}{The method signature}
}
\description{
Retrieve the next applicable method after the current one
}
