% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method_register}
\alias{method_register}
\title{Register R7 methods}
\usage{
method_register()
}
\description{
When registering methods for R7 generics defined in other packages you must
put \code{method_register()} in your packages \link{.onLoad} function.
}
