% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{methods_register}
\alias{methods_register}
\alias{S7_on_build}
\title{Package hooks for S7 methods}
\usage{
methods_register()

S7_on_build()
}
\value{
Nothing; both functions are called for their side-effects.
}
\description{
When using S7 in a package, add two hooks to your \code{zzz.R}:
\itemize{
\item Call \code{methods_register()} from \code{.onLoad()}. This is S7's way of
registering methods, rather than using \code{NAMESPACE} directives like S3 and
S4 do. It ensures that methods for generics (S3, S4, and S7) defined in
other packages are registered as needed when your package is loaded. This
is only strictly necessary if you register methods for generics in other
packages, but there's no harm in always including it and it ensures you
won't forget later.
\item Call \code{S7_on_build()} at the top level (i.e. \emph{not} inside \code{.onLoad()}).
This avoids embedding copies of external, S3, and S4 generics in your
package when you use \verb{method<-}.
}

See \code{vignette("packages")} for more details.
}
\examples{
# In zzz.R:
.onLoad <- function(...) {
  S7::methods_register()
}
S7::S7_on_build()
}
