% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\docType{data}
\name{missing_class}
\alias{missing_class}
\alias{any_class}
\title{Special dispatch types}
\format{
An object of class \code{R7_missing} of length 0.

An object of class \code{R7_any} of length 0.
}
\usage{
missing_class

any_class
}
\description{
\itemize{
\item Use \code{missing_class} when the user has not supplied an argument
\item Use \code{any_class} for a default method that is called only if no other
methods are matched
}
}
\examples{
foo <- new_generic("foo", "x")
method(foo, integer) <- function(x) "integer"
method(foo, missing_class) <- function(x) "missing"
method(foo, any_class) <- function(x) "fallback"

foo(1)
foo()
foo("x")
}
\keyword{datasets}
