% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{new_union}
\alias{new_union}
\title{Define a class union}
\usage{
new_union(...)
}
\arguments{
\item{...}{The classes to include in the union, either looked up by named or
by passing the \code{R7_class} objects directly.}
}
\description{
A class union represents a list of possible classes. It is used in
properties to allow a property to be one of a set of classes, and in method
dispatch as a convenience for defining a method for multiple classes.
}
