% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-dispatch.R
\name{next_method}
\alias{next_method}
\title{Retrieve the next applicable method after the current one}
\usage{
next_method()
}
\description{
Retrieve the next applicable method after the current one
}
\keyword{internal}
