% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lgl}
\alias{lgl}
\alias{int}
\alias{dbl}
\alias{num}
\alias{cpl}
\alias{chr}
\alias{fct}
\alias{raw}
\alias{fn}
\alias{lst}
\alias{df}
\alias{env}
\title{R7 base classes}
\usage{
lgl(x = logical())

int(x = integer())

dbl(x = double())

num(x = numeric())

cpl(x = complex())

chr(x = character())

fct(x = factor())

raw(x = raw())

fn(x = function() NULL)

lst(x = list())

df(x = data.frame())

env(x = emptyenv())
}
\arguments{
\item{x}{default value for the object.}
}
\description{
R7 base classes
}
