% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{R7_data}
\alias{R7_data}
\alias{R7_data<-}
\title{Get/set underlying "base" data}
\usage{
R7_data(object)

R7_data(object, check = TRUE) <- value
}
\arguments{
\item{object}{An object from a R7 class}

\item{check}{If \code{TRUE}, check that \code{value} is of the correct type and run
\code{\link[=validate]{validate()}} on the object before returning.}

\item{value}{Object used to replace the underlying data.}
}
\description{
When an R7 class inherits from an existing base type, it can be useful
to work with the underlying object, i.e. the R7 object stripped of class
and all properties.
}
\examples{
text <- new_class("text", parent = class_character)
y <- text(c(foo = "bar"))
y
R7_data(y)

R7_data(y) <- c("a", "b")
y
}
