% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{r7_generic}
\alias{r7_generic}
\alias{r7_method}
\title{r7 generics and method objects}
\usage{
r7_generic(name, signature, fun)

r7_method(generic, signature, fun)
}
\arguments{
\item{name, generic}{The name or generic object of the generic}

\item{signature}{The signature of the generic}

\item{fun}{The function to use as the body of the generic.}
}
\description{
r7 generics and method objects
}
