% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio.R, R/radio2.R
\name{radio_ui}
\alias{radio_ui}
\title{Framework 7 Radio Button}
\usage{
radio_ui(id, ...)

radio_ui(id, ...)
}
\arguments{
\item{id}{the namespace id of the radio UI}

\item{...}{parameters passed to the uiOutput}
}
\description{
The framework 7 radio button input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs

The framework 7 radio button input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
