% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spec.R
\name{S7_class_desc}
\alias{S7_class_desc}
\title{Format a class specification as a string}
\usage{
S7_class_desc(class)
}
\arguments{
\item{x}{A class specification, processed with \code{\link[=as_class]{as_class()}}.}
}
\value{
A string.
}
\description{
\code{S7_class_desc()} turns any class specification accepted by \code{\link[=as_class]{as_class()}}
into a short human-readable string, suitable for use in error messages or
}
\examples{
S7_class_desc(class_integer)
S7_class_desc(new_S3_class("data.frame"))
S7_class_desc(class_integer | class_double)
S7_class_desc(NULL)
}
