% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-group.R
\name{S7_group_generics}
\alias{S7_group_generics}
\alias{S7_Math}
\alias{S7_Ops}
\alias{S7_Complex}
\alias{S7_Summary}
\alias{find_base_generic}
\title{S7 Group Generics}
\usage{
S7_Math(x, ..., .Generic)

S7_Ops(e1, e2, ..., .Generic)

S7_Complex(z, ..., .Generic)

S7_Summary(x, ..., na.rm = FALSE, .Generic)

find_base_generic(.Generic)
}
\arguments{
\item{x, z, e1, e2}{Objects used for dispatch.}

\item{..., na.rm}{Additional arguments passed to methods.}

\item{.Generic}{The name of the generic being dispatched on, i.e. if you've
defined a method for \code{S7_Math} and the user calls \code{abs()} then \code{.Generic}
will be \code{"abs"}.

Use \code{find_base_generic()} to find the base generic that corresponds to the
generic name.}
}
\description{
Group generics allow you to implement methods for many generics at once.
You cannot call a group generic directly; instead it is called automatically
by members of the group if a more specific method is not found. For example,
if you define a method for the \code{S7_Math} group generic, it will be called
when you call \code{abs()}, \code{sign()}, \code{sqrt()}, and many other similar generics
(see below for a complete list).
}
\section{Methods}{
The group generics contain the following methods:
\itemize{
\item \code{Ops}: \code{+}, \code{-}, \code{*}, \code{^}, \code{\%\%}, \code{\%/\%}, \code{/}, \code{==}, \code{>}, \code{<}, \code{!=}, \code{<=}, \code{>=}, \code{&}, \code{|}
\item \code{Math}: \code{abs}, \code{sign}, \code{sqrt}, \code{ceiling}, \code{floor}, \code{trunc}, \code{cummax}, \code{cummin}, \code{cumprod}, \code{cumsum}, \code{exp}, \code{expm1}, \code{log}, \code{log10}, \code{log2}, \code{log1p}, \code{cos}, \code{cosh}, \code{sin}, \code{sinh}, \code{tan}, \code{tanh}, \code{acos}, \code{acosh}, \code{asin}, \code{asinh}, \code{atan}, \code{atanh}, \code{cospi}, \code{sinpi}, \code{tanpi}, \code{gamma}, \code{lgamma}, \code{digamma}, \code{trigamma}, \code{round}, \code{signif}
\item \code{Summary}: \code{max}, \code{min}, \code{range}, \code{prod}, \code{sum}, \code{any}, \code{all}
\item \code{Complex}: \code{Arg}, \code{Conj}, \code{Im}, \code{Mod}, \code{Re}
\item \code{matrixOps}: \code{\%*\%}
}
}

