% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherits.R
\name{S7_inherits}
\alias{S7_inherits}
\alias{check_is_S7}
\title{Does this object inherit from a class?}
\usage{
S7_inherits(x, class = NULL)

check_is_S7(
  x,
  class = NULL,
  arg = deparse(substitute(x)),
  call = sys.call(-1L)
)
}
\arguments{
\item{x}{An object.}

\item{class}{A class specification (anything accepted by \code{\link[=as_class]{as_class()}}:
an S7 class, S7 union, S3 class, S4 class, base type wrapper like
\link{class_integer}, or \link{class_any}/\link{class_missing}). If \code{NULL}, only tests
whether \code{x} is an S7 object, without testing for a specific class.}

\item{arg}{Argument name used in error message.}

\item{call}{The call to report in the error message. Defaults to the
calling function.}
}
\value{
\itemize{
\item \code{S7_inherits()} returns a single \code{TRUE} or \code{FALSE}.
\item \code{check_is_S7()} returns nothing; it's called for its side-effects.
}
}
\description{
\itemize{
\item \code{S7_inherits()} returns \code{TRUE} or \code{FALSE}.
\item \code{check_is_S7()} throws an error if \code{x} isn't the specified \code{class}.
}
}
\note{
Starting with \R 4.3.0, \code{base::inherits()} can accept an S7 class as
the second argument, supporting usage like \code{inherits(x, Foo)}.
}
\examples{
Foo1 := new_class()
Foo2 := new_class()

S7_inherits(Foo1(), Foo1)
check_is_S7(Foo1())
check_is_S7(Foo1(), Foo1)

S7_inherits(Foo1(), Foo2)
try(check_is_S7(Foo1(), Foo2))

# Also works with other class specifications
S7_inherits(1L, class_integer)
S7_inherits(data.frame(), new_S3_class("data.frame"))
S7_inherits(1L, class_integer | class_character)

if (getRversion() >= "4.3.0") {
  inherits(Foo1(), Foo1)
}
}
