% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_server}
\alias{select_server}
\title{Framework 7 select server-side logic}
\usage{
select_server(
  input,
  output,
  session,
  label,
  choices,
  selected = NULL,
  width = NULL
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{label}{The label on the select UI}

\item{choices}{The choices available for the select UI}

\item{selected}{The option selected by default for the select UI}

\item{width}{Width of the select UI}
}
\description{
The server-side function used in
shinyMobile developed shiny applications
}
