% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_ui}
\alias{select_ui}
\title{Framework 7 select}
\usage{
select_ui(id, container = div, class = "")
}
\arguments{
\item{id}{the namespace id of the select UI}

\item{container}{the container wrapper for the select UI}

\item{class}{the CSS class of the select UI rendered, default block}
}
\description{
The framework 7 select input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
