% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider_server}
\alias{slider_server}
\title{Slider declaration for the server-side function}
\usage{
slider_server(
  input,
  output,
  session,
  label_ = shiny::h4("Weeks"),
  min_ = 1,
  max_ = 26,
  value_ = 3,
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{label_}{The title of the slider object}

\item{min_}{The minimum set for the slider}

\item{max_}{The maximum set for the slider}

\item{value_}{The default value set for the UI}

\item{...}{Additional fields to be passed into the \code{shinyMobile::f7Slider()}}
}
\description{
THe server-side function for the slider that is established a
`shinyMobile` application
}
