% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_select2.R
\name{smart_select2_server}
\alias{smart_select2_server}
\title{Framework 7 smart select server-side logic}
\usage{
smart_select2_server(input, output, session, ...)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{...}{Additional parameters for f7SmartSelect()}
}
\description{
The server-side function used in
shinyMobile developed shiny applications
}
