% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_select.R
\name{smart_select_server}
\alias{smart_select_server}
\title{Framework 7 smart select server-side logic}
\usage{
smart_select_server(
  input,
  output,
  session,
  label,
  choices,
  selected = NULL,
  multiple = F,
  width = NULL,
  openIn = c("page", "sheet", "popup", "popover"),
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{label}{The label on the select UI}

\item{choices}{The choices available for the select UI}

\item{selected}{The option selected by default for the select UI}

\item{multiple}{Logical. Enable multiple selection of choices}

\item{width}{Width of the select UI}

\item{openIn}{option for the method of opening smart select}

\item{...}{Additional parameters for f7SmartSelect()}
}
\description{
The server-side function used in
shinyMobile developed shiny applications
}
