% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_select.R
\name{smart_select_ui}
\alias{smart_select_ui}
\title{Framework 7 select}
\usage{
smart_select_ui(id)
}
\arguments{
\item{id}{the namespace id of the select UI}
}
\description{
The framework 7 select input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
