% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{stepper_server}
\alias{stepper_server}
\title{Framework 7 stepper server-side logic}
\usage{
stepper_server(
  input,
  output,
  session,
  label = shiny::h4("Weeks"),
  min_ = 1,
  max_ = 26,
  value_ = 3,
  ...
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{label}{The label on the stepper UI}

\item{min_}{The minimum value of the stepper UI}

\item{max_}{The maximum value of the stepper UI}

\item{value_}{The value selected of the stepper UI}

\item{...}{additional parameters passed to the \code{f7Stepper()} function}
}
\description{
The server-side function used in
shinyMobile developed shiny applications
}
