% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{stepper_ui}
\alias{stepper_ui}
\title{Framework 7 stepper}
\usage{
stepper_ui(id)
}
\arguments{
\item{id}{the namespace id of the stepper UI}
}
\description{
The framework 7 stepper input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
