% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title_ui}
\alias{title_ui}
\title{Title creation for UI render}
\usage{
title_ui(id)
}
\arguments{
\item{id}{The id used to reference the title object}
}
\description{
A UI-side declaration of the title component
using the Framework 7 structure
}
