% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_reactive.R
\name{to_reactive}
\alias{to_reactive}
\title{Check and convert a value into a reactive function}
\usage{
to_reactive(x)
}
\arguments{
\item{x}{object to convert if NOT already a reactive function}
}
\value{
Reactive function
}
\description{
For some use cases in shiny app development, you will need to check if a value is a
reactive function or reactive value.  To preserve the generalizeable functionality of the ui and server functions,
this supporting function is used to convert a static value or values into a reactive declaration to be
used later on in the function.
}
