% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{toggle_server}
\alias{toggle_server}
\title{Toggle declaration for the server-side function}
\usage{
toggle_server(
  input,
  output,
  session,
  label_ = "?",
  checked = TRUE,
  color_ = "blue"
)
}
\arguments{
\item{input}{list of inputs used in the shiny application session}

\item{output}{list of outputs used the shiny application session}

\item{session}{The shiny app session object}

\item{label_}{The title of the toggle object}

\item{checked}{The minimum set for the toggle}

\item{color_}{The maximum set for the toggle}
}
\value{
[reactive] `logical`: `TRUE` or `FALSE`
}
\description{
THe server-side function for the toggle that is established as a
`shinyMobile` application
}
