% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{toggle_ui}
\alias{toggle_ui}
\title{Toggle UI in the Framework 7 structure}
\usage{
toggle_ui(id)
}
\arguments{
\item{id}{The id used to reference the toggle object}
}
\description{
A UI-side declaration of the toggle UI used for selecting
numbers or strings
}
