% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-spec.R
\name{as_class}
\alias{as_class}
\title{Standard class specifications}
\usage{
as_class(x, arg_name = NULL)
}
\arguments{
\item{x}{A class specification.
\itemize{
\item An R7 class object or class union.
\item An S3 class object, created by \code{s3_class()}.
\item An S4 class object.
\item A base type specified either with its constructor (\code{logical}, \code{integer},
\code{double} etc) or its name (\code{"logical"}, \code{"integer"}, "\code{double}" etc).
}}

\item{arg_name}{Argument name used when generating errors.}
}
\description{
This is used as the interface between R7 and R's other OO systems, allowing
you to build R7 classes and methods on top of base types, informal S3
classes, and formal S4 classes.
}
