% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\docType{data}
\name{base_s3_classes}
\alias{base_s3_classes}
\alias{class_factor}
\alias{class_Date}
\alias{class_POSIXct}
\alias{class_data.frame}
\title{S7 wrappers for key S3 classes}
\usage{
class_factor

class_Date

class_POSIXct

class_data.frame
}
\description{
S7 bundles \link[=new_S3_class]{S3 definitions} for key S3 classes provided by
the base package:
\itemize{
\item \code{class_data.frame} for data frames.
\item \code{class_Date} for dates.
\item \code{class_factor} for factors.
\item \code{class_POSIXct} for \code{POSIXct} date-times.
}
}
\keyword{datasets}
