% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{class_get}
\alias{class_get}
\title{Retrieve the R7 class corresponding to a name}
\usage{
class_get(name, envir = parent.frame())
}
\arguments{
\item{name}{The name of the R7 class}

\item{envir}{The environment to look for the name}
}
\description{
Retrieve the R7 class corresponding to a name
}
