% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\docType{data}
\name{class_missing}
\alias{class_missing}
\title{Dispatch on a missing argument}
\usage{
class_missing
}
\value{
Sentinel objects used for special types of dispatch.
}
\description{
Use \code{class_missing} to dispatch when the user has not supplied an argument,
i.e. it's missing in the sense of \code{\link[=missing]{missing()}}, not in the sense of
\code{\link[=is.na]{is.na()}}.
}
\examples{
foo <- new_generic("foo", "x")
method(foo, class_numeric) <- function(x) "number"
method(foo, class_missing) <- function(x) "missing"
method(foo, class_any) <- function(x) "fallback"

foo(1)
foo()
foo("")
}
\keyword{datasets}
