% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert an object from one type to another}
\usage{
convert(from, to, ...)
}
\arguments{
\item{from}{An R7 object to convert.}

\item{to}{An R7 class specification, passed to \code{\link[=as_class]{as_class()}}.}

\item{...}{Other arguments passed to custom \code{convert()} methods.}
}
\description{
\code{convert()} uses double-dispatch, because conversion depends on both \code{from}
and \code{to}. The dispatch is non-standard, because \code{to} is a class (not an
object), and it does not take advantage of inheritance (because if you
convert \code{x} to \code{superFoo} you shouldn't get an instance of \code{Foo} back).

\code{convert()} provides built-in implementations if \code{from} inherits from \code{to}.
This default strips any properties that \code{from} possesses that \code{to} does not,
and resets the class.
}
