% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert an object from one type to another}
\usage{
convert(from, to, ...)
}
\arguments{
\item{from}{An R7 object to convert.}

\item{to}{An R7 class specification, passed to \code{\link[=as_class]{as_class()}}.}

\item{...}{Other arguments passed to custom \code{convert()} methods.}
}
\description{
\code{convert()} is a non-standard generic: it uses double dispatch on the first class of
\code{from} and \code{to} (and unlike normal dispatch, \code{to} is a class, not an object).

\code{convert()} provides an automatic fallback if \code{from} inherits from \code{to}. You
can override this if you need some special behavior other than simply
stripping class.
}
