% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{find_call}
\alias{find_call}
\title{Find a call in expression}
\usage{
find_call(x, call)
}
\arguments{
\item{x}{An language object in which to search for a particular function
call.}

\item{call}{A quoted function name to search for. Both the plain function
name and qualified function name from the package namespace are considered.}
}
\value{
The \code{call} object if found, or \code{NULL} otherwise.
}
\description{
Find a call (namespaced or plain) in an language object.
}
\keyword{internal}
