% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{generic_new}
\alias{generic_new}
\title{Define a new generic}
\usage{
generic_new(name, signature)
}
\arguments{
\item{name}{The name or generic object of the generic}

\item{signature}{The signature of the generic}
}
\description{
Define a new generic
}
