% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method<-}
\alias{method<-}
\title{Register a R7 method for a generic}
\usage{
method(generic, signature) <- value
}
\arguments{
\item{generic}{A generic function.}

\item{signature}{A method signature, a list of R7 class constructors
(produced by \code{\link[=new_class]{new_class()}}) or names of S3 or S4 classes.}

\item{value}{A function that implements the generic specification for the
given \code{signature}. The arguments must be compatible with the generic.}
}
\description{
A generic defines the interface of a function. Once you have created a
generic with \code{\link[=new_generic]{new_generic()}}, you provide implememtnations for specific
signatures by registering methods with \verb{method<-}
}
\examples{
# Create a generic
bizarro <- new_generic("bizarro", dispatch_args = "x")
# Register some methods
method(bizarro, "numeric") <- function(x) rev(x)
method(bizarro, s3_class("data.frame")) <- function(x) {
  x[] <- lapply(x, bizarro)
  rev(x)
}

# Using a generic calls the methods automatically
bizarro(head(mtcars))
}
