% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\alias{new_method}
\alias{method<-}
\title{Retrieve or register an R7 method for a generic}
\usage{
method(generic, signature)

new_method(generic, signature, method, package = NULL)

method(generic, signature) <- value
}
\arguments{
\item{generic}{The generic to retrieve or register}

\item{signature}{The method signature}

\item{method, value}{The new function to use as the method.}

\item{package}{The package to register the method in, only used for soft
dependencies. The default \code{NULL} looks up the package based on the parent
frame.}
}
\description{
Retrieve or register an R7 method for a generic
}
