% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\alias{method_register}
\title{Retrieve or register an r7 method for a generic}
\usage{
method(generic, classes)

method_register(generic, class, value)
}
\arguments{
\item{generic}{The generic to retrieve or register}

\item{class, classes}{The class to register the method for}

\item{value}{The new function to use for the method.}
}
\description{
Retrieve or register an r7 method for a generic
}
