% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\alias{method_register}
\alias{method<-}
\title{Retrieve or register an r7 method for a generic}
\usage{
method(generic, signature)

method_register(generic, signature, value)

method(generic, signature) <- value
}
\arguments{
\item{generic}{The generic to retrieve or register}

\item{signature}{The method signature}

\item{value}{The new function to use for the method.}
}
\description{
Retrieve or register an r7 method for a generic
}
