% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{method}
\alias{method}
\title{Retrieve a method for an R7 generic}
\usage{
method(generic, classes = NULL, objects = NULL)
}
\arguments{
\item{generic}{A generic function, either created by \code{\link[=new_generic]{new_generic()}},
\code{\link[=new_external_generic]{new_external_generic()}}, or an existing S3 generic.}

\item{classes, objects}{Perform introspection either with \code{classes}
(processed with \code{\link[=as_class]{as_class()}}) or a concrete objects.

If \code{generic} does multiple dispatch both \code{objects} and \code{classes} need
to be wrapped in a list.}
}
\value{
A function with class <R7_method>.
}
\description{
\code{method()} takes a generic and signature and retrieves the corresponding
method. This is rarely needed because most of the time you'll rely on the
the generic, via \code{\link[=method_call]{method_call()}}, to find and call the method for you.
However, this introspection is useful if you want to see the implementation
of a specific method.
}
\examples{
# Create a generic and register some methods
bizarro <- new_generic("bizarro", "x")
method(bizarro, "numeric") <- function(x) rev(x)
method(bizarro, S3_class("factor")) <- function(x) {
  levels(x) <- rev(levels(x))
  x
}

# Printing the generic shows the registered method
bizarro

# And you can use method() to inspect specific implementations
method(bizarro, classes = "integer")
method(bizarro, objects = 1)
method(bizarro, classes = S3_class("factor"))
}
