% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method_next}
\alias{method_next}
\title{Retrieve the next applicable method after the current one}
\usage{
method_next(generic, signature, previous_classes = NULL, previous_methods)
}
\description{
Retrieve the next applicable method after the current one
}
