% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-generic.R
\name{methods_register}
\alias{methods_register}
\title{Register methods in a package}
\usage{
methods_register()
}
\description{
When using S7 in a package you must always call \code{methods_register()} when
your package is loaded. This ensures that methods are registered as needed
when packages that provide generics you provide methods for are loaded.
}
\examples{
.onLoad <- function(...) {
  S7::methods_register()
}
}
