% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_class}
\alias{new_class}
\title{Define a new R7 class}
\usage{
new_class(
  name,
  parent = R7_object,
  constructor = function(.data = NULL, ...) new_object(.data, ...),
  validator = function(x) NULL,
  properties = list()
)
}
\arguments{
\item{name}{The name of the class}

\item{parent}{The parent class}

\item{constructor}{The constructor function}

\item{validator}{The validation function}

\item{properties}{A list of properties for the class}
}
\description{
Define a new R7 class
}
