% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_class}
\alias{new_class}
\alias{new_object}
\title{Define a new R7 class}
\usage{
new_class(
  name,
  parent = R7_object,
  package = NULL,
  properties = list(),
  constructor = NULL,
  validator = NULL
)

new_object(.parent, ...)
}
\arguments{
\item{name}{The name of the class, as a string.}

\item{parent}{The parent class to inherit behavior from.
There are four options:
\itemize{
\item The R7 class, like \link{R7_object}.
\item An S3 class wrapped by \code{\link[=new_S3_class]{new_S3_class()}}.
\item A base type, like \code{logical}, \code{double}, or \code{character}.
}}

\item{package}{Package name. It is good practice to set the package
name when exporting an R7 class from a package because it includes
the package name in the class name when it's used for dispatch. This
allows different packages to use the same name to refer to different
classes.}

\item{properties}{A list specifying the properties (data) that
every object of the class will possess. Each property can either be
a named string (specifying the class), or a call to \code{\link[=new_property]{new_property()}},
allowing greater flexibility.}

\item{constructor}{The constructor function. Advanced use only.

A custom constructor should call \code{new_object()} to create the R7 object.
The first argument, \code{.data}, should an instance of the parent class. The
subsequent arguments are used to set the properties.}

\item{validator}{A function taking a single argument, \code{self}, the object
to validate.

The job of a validator is to determine whether the object is valid,
i.e. if the current property values form an allowed combination. The
types of the properties are always automatically validated so the job of
the validator is to verify that the \emph{values} of individual properties are
ok (i.e. maybe a property should have length 1, or should always be
positive), or that the \emph{combination} of values of multiple properties is ok.
It is called after construction and whenever any property is set.

The validator should return \code{NULL} if the object is valid. If not, it
should return a character vector where each element describes a single
problem, using \verb{@prop_name} to describe where the problem lies.

See \code{validate()} for more details, examples, and how to temporarily
suppress validation when needed.}

\item{.parent, ...}{Parent object and named properties used to construct the
object.}
}
\value{
A object constructor, a function that can be used to create objects
of the given class.
}
\description{
A class specifies the properties (data) that each of its objects will
possess. The class, and its parent, determines which method will be used
when an object is passed to a generic.
}
\examples{
# Create an class that represents a range using a numeric start and end
range <- new_class("range",
  properties = list(
    start = "numeric",
    end = "numeric"
  )
)
r <- range(start = 10, end = 20)
r
# get and set properties with @
r@start
r@end <- 40
r@end

# R7 automatically ensures that properties are of the declared types:
try(range(start = "hello", end = 20))

# But we might also want to use a validator to ensure that start and end
# are length 1, and that start is < end
range <- new_class("range",
  properties = list(
    start = "numeric",
    end = "numeric"
  ),
  validator = function(self) {
    if (length(self@start) != 1) {
      "@start must be a single number"
    } else if (length(self@end) != 1) {
      "@end must be a single number"
    } else if (self@end < self@start) {
      "@end must be great than or equal to @start"
    }
  }
)
try(range(start = c(10, 15), end = 20))
try(range(start = 20, end = 10))

r <- range(start = 10, end = 20)
try(r@start <- 25)
}
