% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{new_generic}
\alias{new_generic}
\title{Define a new generic}
\usage{
new_generic(name, signature = NULL, fun = NULL)
}
\arguments{
\item{name}{The name or generic object of the generic}

\item{signature}{The signature of the generic}

\item{fun}{The function to use as the body of the generic.}
}
\description{
Define a new generic
}
