% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{prop}
\alias{prop}
\alias{prop<-}
\alias{@.R7_object}
\alias{@}
\title{Get/set a property}
\usage{
prop(object, name)

prop(object, name, check = TRUE) <- value

object@name
}
\arguments{
\item{object}{An object from a R7 class}

\item{name}{The name of the parameter as a character. Partial matching
is not performed.}

\item{check}{If \code{TRUE}, check that \code{value} is of the correct type and run
\code{\link[=validate]{validate()}} on the object before returning.}

\item{value}{A new value for the property. The object is automatically
checked for validity after the replacement is done.}
}
\description{
\itemize{
\item \code{prop(x, "name")} / \code{prop@name} get the value of the a property,
erroring if it the property doesn't exist.
\item \code{prop(x, "name") <- value} / \code{prop@name <- value} set the value of
a property.
}
}
\examples{
horse <- new_class("horse", properties = list(
  name = class_character,
  colour = class_character,
  height = class_numeric
))
lexington <- horse(colour = "bay", height = 15, name = "Lex")
lexington@colour
prop(lexington, "colour")

lexington@height <- 14
prop(lexington, "height") <- 15
}
