% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{prop}
\alias{prop}
\alias{prop<-}
\alias{@}
\alias{@<-.object}
\title{Extract or replace a property}
\usage{
prop(obj, name)

prop(obj, name) <- value

`@`(obj, name)

\method{@}{object}(obj, name) <- value
}
\arguments{
\item{obj}{An object from a R7 class}

\item{name}{The name of the parameter as a character. No partial matching is done.}

\item{value}{A replacement value for the parameter. The object is
automatically checked for validity after the replacement is done.}
}
\description{
Extract or replace a property
}
