% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{prop}
\alias{prop}
\alias{prop_safely}
\alias{prop_names}
\alias{props}
\alias{prop_exists}
\alias{prop<-}
\alias{@}
\title{Get or set value of a property}
\usage{
prop(object, name)

prop_safely(object, name)

prop_names(object)

props(object)

prop_exists(object, name)

prop(object, name, check = TRUE) <- value

object@name
}
\arguments{
\item{object}{An object from a R7 class}

\item{name}{The name of the parameter as a character. Partial matching
is not performed.}

\item{check}{If \code{TRUE}, check that \code{value} is of the correct type and run
\code{\link[=validate]{validate()}} on the object before returning.}

\item{value}{A replacement value for the parameter. The object is
automatically checked for validity after the replacement is done.}
}
\description{
\itemize{
\item \code{prop()} and \code{@}, gets the value of the given property, throwing an
error if the property doesn't exist for that object.
\item \code{prop_safely()} returns \code{NULL} if a property doesn't exist,
rather than throwing an error.
\item \verb{prop<-} and \verb{@<-} set a new value for the given property.
\item \code{props()} returns a list of all properties
\item \code{prop_names()} returns the names of the properties
\item \code{prop_exists(x, "prop")} returns \code{TRUE} iif \code{x} has property \code{prop}.
}
}
\examples{
horse <- new_class("horse", properties = list(
  name = "character",
  colour = "character",
  height = "numeric"
))
lexington <- horse(colour = "bay", height = 15, name = "Lex")
lexington@colour
prop(lexington, "colour")

lexington@height <- 14
prop(lexington, "height") <- 15

try(prop(lexington, "age"))
prop_safely(lexington, "age")
}
