% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property}
\alias{property}
\alias{property<-}
\alias{@}
\title{Extract or replace a property}
\usage{
property(obj, name)

property(obj, name) <- value

obj@name
}
\arguments{
\item{obj}{An object from a R7 class}

\item{name}{The name of the parameter as a character. No partial matching is done.}

\item{value}{A replacement value for the parameter. The object is
automatically checked for validity after the replacement is done.}
}
\description{
Extract or replace a property
}
