% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property_new}
\alias{property_new}
\title{Define a new property}
\usage{
property_new(name, class = NULL, accessor = NULL)
}
\arguments{
\item{name}{The name of the property}

\item{class}{The class of the property}

\item{accessor}{The accessor use to retrieve the property (if any)}
}
\description{
Define a new property
}
