% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{props}
\alias{props}
\alias{props<-}
\title{Get/set multiple properties}
\usage{
props(object)

props(object) <- value
}
\arguments{
\item{object}{An object from a R7 class}

\item{value}{A named list of values. The object is checked for validity
only after all replacements are performed.}
}
\description{
\itemize{
\item \code{props(x)} returns all properties.
\item \code{props(x) <- list(name1 = val1, name2 = val2)} sets multiple properties.
}
}
\examples{
horse <- new_class("horse", properties = list(
  name = "character",
  colour = "character",
  height = "numeric"
))
lexington <- horse(colour = "bay", height = 15, name = "Lex")

props(lexington)
props(lexington) <- list(height = 14, name = "Lexigonton")
lexington
}
