% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{r7_data}
\alias{r7_data}
\alias{r7_data<-}
\title{Get/set underlying "base" data}
\usage{
r7_data(object)

r7_data(object, check = TRUE) <- value
}
\arguments{
\item{object}{An object from a R7 class}

\item{check}{If \code{TRUE}, check that \code{value} is of the correct type and run
\code{\link[=validate]{validate()}} on the object before returning.}

\item{value}{A replacement value for the parameter. The object is
automatically checked for validity after the replacement is done.}
}
\description{
When an R7 class inherits from an existing base type, it can be useful
to work with the underlying object, i.e. the R7 object stripped of class
and all properties.
}
\examples{
text <- new_class("text", parent = "character")
y <- text(c(foo = "bar"))
str(r7_data(y))
}
