% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{r7_generic}
\alias{r7_generic}
\alias{r7_method}
\title{r7 generics and method objects}
\usage{
r7_generic(name, signature, envir = parent.frame())

r7_method(generic, signature, fun)
}
\arguments{
\item{name, generic}{The name or generic object of the generic}

\item{signature}{The signature of the generic}

\item{envir}{The environment to lookup the generic in.}

\item{fun}{The function of the method'}
}
\description{
r7 generics and method objects
}
