% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{R7_generic}
\alias{R7_generic}
\alias{R7_method}
\title{R7 generics and method objects}
\usage{
R7_generic(name, signature, fun)

R7_method(generic, signature, fun)
}
\arguments{
\item{name, generic}{The name or generic object of the generic}

\item{signature}{The signature of the generic}

\item{fun}{The function to use as the body of the generic.}
}
\description{
R7 generics and method objects
}
