% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{S7_class}
\alias{S7_class}
\title{Retrieve the S7 class of an object}
\usage{
S7_class(object)
}
\arguments{
\item{object}{The S7 object}
}
\value{
An \link[=new_class]{S7 class}.
}
\description{
Given an S7 object, find it's class.
}
\examples{
foo <- new_class("foo")
S7_class(foo())
}
