% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{S7_object}
\alias{S7_object}
\title{Base S7 class}
\usage{
S7_object()
}
\value{
The base S7 object.
}
\description{
The base class from which all S7 classes eventually inherit from.
}
\examples{

S7_object
}
\keyword{internal}
