% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_ui}
\alias{select_ui}
\title{Framework 7 select}
\usage{
select_ui(id)
}
\arguments{
\item{id}{the namespace id of the select UI}
}
\description{
The framework 7 select input that
relies on the shinyMobile package to develop
a compatible UI useful in PWAs
}
