% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider_ui}
\alias{slider_ui}
\title{Slider UI in the Framework 7 structure}
\usage{
slider_ui(id)
}
\arguments{
\item{id}{The id used to reference the slider object}
}
\description{
A UI-side declaration of the slider UI used for selecting
numbers or strings
}
