% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validate}
\alias{validate}
\alias{valid_eventually}
\alias{valid_implicitly}
\title{Validation of R7 objects}
\usage{
validate(object)

valid_eventually(object, fun)

valid_implicitly(object, fun)
}
\arguments{
\item{object}{An R7 object}

\item{fun}{A function to call on the object before validation.}
}
\description{
\code{\link[=validate]{validate()}} calls the validation of an R7 object. This is done
automatically when creating new objects (at the end of \link{new_object}) and
when setting any property.
}
\details{
\code{\link[=valid_eventually]{valid_eventually()}} disables validation of properties, runs a function on
the object, then validates the object.

\code{\link[=valid_implicitly]{valid_implicitly()}} does the same but does not validate the object at the end.

\code{\link[=valid_implicitly]{valid_implicitly()}} should only be used rarely in performance critical code
where you are certain a sequence of operations cannot produce an invalid
object.
}
